requirements.txt
absl-py==0.14.1
aiohttp==3.7.4.post0
aiohttp-retry==2.4.6
appdirs==1.4.4
astunparse==1.6.3
async-timeout==3.0.1
atpublic==2.3
attrs==21.2.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
clang==5.0
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
contextvars==2.4
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
dictdiffer==0.9.0
diskcache==5.2.1
distro==1.6.0
dpath==2.0.5
dulwich==0.20.25
dvc==2.8.1
flatbuffers==1.12
flatten-dict==0.4.2
flufl.lock==3.2
fsspec==2021.10.0
ftfy==6.0.3
funcy==1.16
future==0.18.2
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grandalf==0.6
graphviz==0.19.1
grpcio==1.41.0
h5py==3.1.0
idna==3.2
idna-ssl==1.1.0
immutables==0.16
importlib-metadata==4.8.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
mailchecker==4.0.12
Markdown==3.3.4
matplotlib==3.3.4
multidict==5.2.0
nanotime==0.5.2
networkx==2.5.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pathspec==0.8.1
phonenumbers==8.12.34
Pillow==8.3.2
ply==3.11
protobuf==3.18.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
pydot-ng==2.0.0
pydotplus==2.0.2
pygit2==1.6.1
Pygments==2.10.0
pygtrie==2.4.2
pyparsing==3.0.7
python-benedict==0.24.3
python-dateutil==2.8.2
python-fsutil==0.5.0
python-slugify==5.0.2
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rich==10.12.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scipy==1.5.4
shortuuid==1.0.1
shtab==1.4.1
six==1.15.0
smmap==4.0.0
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.62.3
typing-extensions==3.7.4.3
urllib3==1.26.7
voluptuous==0.12.2
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.12.1
xmltodict==0.12.0
yarl==1.7.0
zc.lockfile==2.0
zipp==3.6.0
